refactor(decompose): split packages/libraries/graph-tools/src/commands/folders/folderMaterialize.ts into folder
